+++ lib/src/model/index.dart
@@ -4,3 +4,4 @@ export 'flag.dart';
 export 'flag_and_traits.dart';
 export 'flag_type.dart';
 export 'trait.dart';
+export 'loading.dart';