+++ test/flag_test.dart
@@ -1,3 +1,4 @@
+import 'package:flagsmith/src/model/flag_type.dart';
 import 'package:flagsmith/src/model/flag.dart';
 import 'package:flutter_test/flutter_test.dart';
 
@@ -29,5 +30,34 @@ void main() {
     expect(flag.feature.name, isNotNull);
     expect(flag.feature.type, isNotNull);
     expect(flag.feature.description, isNotNull);
+  });
+
+  test('flag seed - enabled', () {
+    var flagDefault = Flag.seed('feature');
+
+    expect(flagDefault.enabled, true);
+    expect(flagDefault.feature, isNotNull);
+    expect(flagDefault.feature.type, FlagType.flag);
+
+    var flag = Flag.seed('feature');
+
+    expect(flag.enabled, true);
+    expect(flag.feature, isNotNull);
+    expect(flag.feature.type, FlagType.flag);
+  });
+  test('flag seed - disabled', () {
+    var flag = Flag.seed('feature', enabled: false);
+    expect(flag.enabled, false);
+    expect(flag.feature, isNotNull);
+    expect(flag.feature.type, FlagType.flag);
+  });
+
+  test('flag seed - cofig', () {
+    var flag = Flag.seed('feature', enabled: false, value: '1.0.0');
+    expect(flag.enabled, false);
+    expect(flag.feature, isNotNull);
+    expect(flag.feature.type, FlagType.config);
+    expect(flag.stateValue, isNotNull);
+    expect(flag.stateValue, '1.0.0');
   });
 }